+++ R/splr_matrix_class.R
@@ -41,6 +41,8 @@ setGeneric(
   name = "splr",
   def = function(x, a = NULL, b = NULL, rank = NULL,
   dimnames = list(NULL, NULL), ...) {
+
+  standardGeneric("splr")
   #
   # x+ab'
   # x is mxn
@@ -554,6 +556,7 @@ setMethod("norm", signature(x ="splrMatrix", type ="character"),
 #' @rdname innerproduct
 #' @export
 setGeneric("innerproduct", function(x, y){
+  standardGeneric("innerproduct")
   sum(x * y)
 })