+++ NEWS.md
@@ -1,3 +1,9 @@
+
+# iGraphMatch 1.1
+
+* Implemented `graphMatch` class
+* `get_perm` is now called `make_perm` and `get_perm` is a function that take a `graphMatch` object
+
 # iGraphMatch 1.0.1
 
 * Added a `NEWS.md` file to track changes to the package.